Refactory cyrptography interfaces and classes
