+++ src/main/java/org/javalaboratories/core/cryptography/FileCryptographyResultImpl.java
@@ -15,8 +15,6 @@
  */
 package org.javalaboratories.core.cryptography;
 
-import org.javalaboratories.core.cryptography.keys.CryptographyResultImpl;
-
 import java.io.File;
 import java.security.Key;
 import java.util.Objects;
